Issue #42: Rename environment directories
