+++ bsk_rl/envs/__init__.py
@@ -1,14 +1,14 @@
-from bsk_rl.envs.AgileEOS.gym_env import AgileEOS as AgileEOS
-from bsk_rl.envs.GeneralSatelliteTasking.gym_env import (
+from bsk_rl.envs.agile_eos.gym_env import AgileEOS as AgileEOS
+from bsk_rl.envs.general_satellite_tasking.gym_env import (
     GeneralSatelliteTasking as GeneralSatelliteTasking,
 )
-from bsk_rl.envs.GeneralSatelliteTasking.gym_env import (
+from bsk_rl.envs.general_satellite_tasking.gym_env import (
     SingleSatelliteTasking as SingleSatelliteTasking,
 )
-from bsk_rl.envs.MultiSatAgileEOS.gym_env import MultiSatAgileEOS as MultiSatAgileEOS
-from bsk_rl.envs.MultiSensorEOS.gym_env import MultiSensorEOS as MultiSensorEOS
-from bsk_rl.envs.SimpleEOS.gym_env import SimpleEOS as SimpleEOS
-from bsk_rl.envs.SmallBodyScience.gym_env import SmallBodyScience as SmallBodyScience
-from bsk_rl.envs.SmallBodySciencePOMDP.gym_env import (
+from bsk_rl.envs.multisat_agile_eos.gym_env import MultiSatAgileEOS as MultiSatAgileEOS
+from bsk_rl.envs.multisensor_eos.gym_env import MultiSensorEOS as MultiSensorEOS
+from bsk_rl.envs.simple_eos.gym_env import SimpleEOS as SimpleEOS
+from bsk_rl.envs.small_body_science.gym_env import SmallBodyScience as SmallBodyScience
+from bsk_rl.envs.small_body_science_pomdp.gym_env import (
     SmallBodySciencePOMDP as SmallBodySciencePOMDP,
 )
