+++ examples/general_satellite_tasking/multisat_aeos.py
@@ -1,12 +1,12 @@
 import gymnasium as gym
 import numpy as np
 
-from bsk_rl.envs.GeneralSatelliteTasking.scenario import communication, data
-from bsk_rl.envs.GeneralSatelliteTasking.scenario import satellites as sats
-from bsk_rl.envs.GeneralSatelliteTasking.scenario.environment_features import (
+from bsk_rl.envs.general_satellite_tasking.scenario import communication, data
+from bsk_rl.envs.general_satellite_tasking.scenario import satellites as sats
+from bsk_rl.envs.general_satellite_tasking.scenario.environment_features import (
     CityTargets,
 )
-from bsk_rl.envs.GeneralSatelliteTasking.simulation import environment
+from bsk_rl.envs.general_satellite_tasking.simulation import environment
 from bsk_rl.utilities.initial_conditions import leo_orbit
 
 # This script demonstrates the configuration of an environment with multiple imaging
